Move Individual class to new files
