+++ Pv_mod/Params.cpp
@@ -19,8 +19,12 @@
 #include <cmath>
 
 
-/////////////////////////////////////
-// Misc. functions
+////////////////////////////////////////////////////////////
+//                                                        //
+//  Function declarations                                 //
+//                                                        //
+////////////////////////////////////////////////////////////
+
 double gammln(const double xx);
